improving delete brand

+++ app/src/main/java/com/example/gestordeperfumaria/MainActivity.kt
@@ -4,6 +4,7 @@ import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
+import android.widget.Toast
 import androidx.lifecycle.LiveData
 import androidx.room.Room
 import com.example.gestordeperfumaria.databinding.ActivityMainBinding
@@ -29,24 +30,25 @@ class MainActivity : AppCompatActivity() {
             applicationContext,
             AppDataBase::class.java,
             "db-perfumery"
-        )
+        ).fallbackToDestructiveMigration()
             .allowMainThreadQueries().build()
 
         binding.btnRegister.setOnClickListener {
-            //dbInsertBrand("Avon", 0.2f)
-            dbInsertCosmetic("Perfume", "Natura", 100.0f)
+            dbInsertBrand("Avon", 0.2f)
+            //dbInsertCosmetic("Perfume", 1, 100.0f)
             startActivity(Intent(this, RegisterActivity::class.java))
         }
 
         binding.btnMonitor.setOnClickListener {
             //dbShowAllBrands()
             //dbShowBrand(23)
-            //dbDeleteBrand(10)
+            dbDeleteBrand(3)
             //dbDeleteAllBrands()
             //dbUpdateBrand("Boticario", 0.15f, 25)
             //dbShowAllCosmetics()
             //dbDeleteCosmetic(2)
-            dbUpdateCosmetic("Colonia", "Avon", 25.0f, 6)
+            //dbDeleteAllCosmetics()
+            //dbUpdateCosmetic("Colonia", 2, 25.0f, 6)
             startActivity(Intent(this, MonitorActivity::class.java))
         }
     }
@@ -56,7 +58,8 @@ class MainActivity : AppCompatActivity() {
             val brandEntity = BrandEntity(0, name, profit)
             val result = db.brandDAO.insert(brandEntity)
             if(result > 0) {
-                Log.i(result.toString(), "Marca cadastrada com sucesso!")
+                val message = "Marca cadastrada com sucesso!"
+                Toast.makeText(this@MainActivity, message, Toast.LENGTH_SHORT).show()
             }
         }
         insertBrands.join()
@@ -73,6 +76,14 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun dbDeleteBrand(id: Long) = runBlocking {
+        val listCosmetics = async { db.cosmeticDAO.getAll() }.await()
+        listCosmetics.forEach {
+            if(it.idBrand == id) {
+                val message = "Não é possível excluir a marca selecionada, pois existem cosméticos cadastrados dessa marca!"
+                Toast.makeText(this@MainActivity, message, Toast.LENGTH_SHORT).show()
+                return@runBlocking
+            }
+        }
         val deleteBrand = launch {
             db.brandDAO.delete(id)
         }
@@ -80,10 +91,16 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun dbDeleteAllBrands() = runBlocking {
-        val deleteAllBrands = launch {
-            db.brandDAO.deleteAll()
+        val listCosmetics = db.cosmeticDAO.getAll()
+        if(listCosmetics.isNotEmpty()) {
+            val message = "Ainda existem cosméticos cadastrados, não é possível excluir todas as marcas."
+            Toast.makeText(this@MainActivity, message, Toast.LENGTH_SHORT).show()
+        } else {
+            val deleteAllBrands = launch {
+                db.brandDAO.deleteAll()
+            }
+            deleteAllBrands.join()
         }
-        deleteAllBrands.join()
     }
 
     private fun dbUpdateBrand(name: String, profit: Float, id: Long) = runBlocking {
@@ -93,14 +110,15 @@ class MainActivity : AppCompatActivity() {
         updateBrand.join()
     }
 
-    private fun dbInsertCosmetic(name: String, nameBrand: String, price: Float) = runBlocking {
+    private fun dbInsertCosmetic(name: String, idBrand: Long, price: Float) = runBlocking {
         val insertCosmetic = launch {
             val dateFormat = SimpleDateFormat("dd/MM/yyyy")
             val date = dateFormat.format(Date()).toString()
-            val cosmeticEntity = CosmeticEntity(0, name, nameBrand, date, price)
+            val cosmeticEntity = CosmeticEntity(0, name, idBrand, date, price)
             val result = db.cosmeticDAO.insert(cosmeticEntity)
             if(result > 0) {
-                Log.i(result.toString(), "Cosmético cadastrado com sucesso!")
+                val message = "Cosmético cadastrado com sucesso!"
+                Toast.makeText(this@MainActivity, message, Toast.LENGTH_SHORT).show()
             }
         }
         insertCosmetic.join()
@@ -119,19 +137,15 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun dbDeleteAllCosmetics() = runBlocking {
-        val brands = db.brandDAO.getAll()
-        brands.forEach {
-            it.name
-        }
         val deleteAllCosmetics = launch {
             db.cosmeticDAO.deleteAll()
         }
         deleteAllCosmetics.join()
     }
 
-    private fun dbUpdateCosmetic(name: String, nameBrand: String, price: Float, id: Long) = runBlocking {
+    private fun dbUpdateCosmetic(name: String, idBrand: Long, price: Float, id: Long) = runBlocking {
         val updateCosmetic = launch {
-            db.cosmeticDAO.update(name, nameBrand, price, id)
+            db.cosmeticDAO.update(name, idBrand, price, id)
         }
         updateCosmetic.join()
     }
